Move send in libs/common
