+++ libs/common/src/abstractions/search.service.ts
@@ -1,4 +1,4 @@
-import { SendView } from "../models/view/send.view";
+import { SendView } from "../tools/send/models/view/send.view";
 import { CipherView } from "../vault/models/view/cipher.view";
 
 export abstract class SearchService {
