+++ libs/common/src/tools/send/models/api/send-file.api.ts
@@ -1,4 +1,4 @@
-import { BaseResponse } from "../response/base.response";
+import { BaseResponse } from "../../../../models/response/base.response";
 
 export class SendFileApi extends BaseResponse {
   id: string;
