+++ libs/common/src/tools/send/models/api/send-text.api.ts
@@ -1,4 +1,4 @@
-import { BaseResponse } from "../response/base.response";
+import { BaseResponse } from "../../../../models/response/base.response";
 
 export class SendTextApi extends BaseResponse {
   text: string;
