+++ libs/common/src/tools/send/models/data/send-text.data.ts
