+++ libs/common/src/tools/send/models/domain/send-access.spec.ts
@@ -1,12 +1,12 @@
 // eslint-disable-next-line no-restricted-imports
 import { Substitute, Arg } from "@fluffy-spoon/substitute";
 
-import { SendType } from "@bitwarden/common/enums/sendType";
-import { SendAccess } from "@bitwarden/common/models/domain/send-access";
-import { SendText } from "@bitwarden/common/models/domain/send-text";
-import { SendAccessResponse } from "@bitwarden/common/models/response/send-access.response";
+import { mockEnc } from "../../../../../spec/utils";
+import { SendType } from "../../enums/send-type";
+import { SendAccessResponse } from "../response/send-access.response";
 
-import { mockEnc } from "../../utils";
+import { SendAccess } from "./send-access";
+import { SendText } from "./send-text";
 
 describe("SendAccess", () => {
   let request: SendAccessResponse;
