+++ libs/common/src/tools/send/models/domain/send-file.ts
@@ -1,11 +1,10 @@
 import { Jsonify } from "type-fest";
 
+import Domain from "../../../../models/domain/domain-base";
+import { EncString } from "../../../../models/domain/enc-string";
+import { SymmetricCryptoKey } from "../../../../models/domain/symmetric-crypto-key";
 import { SendFileData } from "../data/send-file.data";
 import { SendFileView } from "../view/send-file.view";
-
-import Domain from "./domain-base";
-import { EncString } from "./enc-string";
-import { SymmetricCryptoKey } from "./symmetric-crypto-key";
 
 export class SendFile extends Domain {
   id: string;
