+++ libs/common/src/tools/send/models/domain/send.spec.ts
@@ -1,16 +1,16 @@
 // eslint-disable-next-line no-restricted-imports
 import { Substitute, Arg, SubstituteOf } from "@fluffy-spoon/substitute";
 
-import { CryptoService } from "@bitwarden/common/abstractions/crypto.service";
-import { EncryptService } from "@bitwarden/common/abstractions/encrypt.service";
-import { SendType } from "@bitwarden/common/enums/sendType";
-import { SendData } from "@bitwarden/common/models/data/send.data";
-import { EncString } from "@bitwarden/common/models/domain/enc-string";
-import { Send } from "@bitwarden/common/models/domain/send";
-import { SendText } from "@bitwarden/common/models/domain/send-text";
-import { ContainerService } from "@bitwarden/common/services/container.service";
+import { makeStaticByteArray, mockEnc } from "../../../../../spec/utils";
+import { CryptoService } from "../../../../abstractions/crypto.service";
+import { EncryptService } from "../../../../abstractions/encrypt.service";
+import { EncString } from "../../../../models/domain/enc-string";
+import { ContainerService } from "../../../../services/container.service";
+import { SendType } from "../../enums/send-type";
+import { SendData } from "../data/send.data";
 
-import { makeStaticByteArray, mockEnc } from "../../utils";
+import { Send } from "./send";
+import { SendText } from "./send-text";
 
 describe("Send", () => {
   let data: SendData;
