+++ libs/common/src/tools/send/models/response/send-access.response.ts
@@ -1,8 +1,7 @@
-import { SendType } from "../../enums/sendType";
+import { BaseResponse } from "../../../../models/response/base.response";
+import { SendType } from "../../enums/send-type";
 import { SendFileApi } from "../api/send-file.api";
 import { SendTextApi } from "../api/send-text.api";
-
-import { BaseResponse } from "./base.response";
 
 export class SendAccessResponse extends BaseResponse {
   id: string;
