+++ libs/common/src/tools/send/models/response/send-file-download-data.response.ts
@@ -1,4 +1,4 @@
-import { BaseResponse } from "./base.response";
+import { BaseResponse } from "../../../../models/response/base.response";
 
 export class SendFileDownloadDataResponse extends BaseResponse {
   id: string = null;
