+++ libs/common/src/tools/send/services/send.service.abstraction.ts
@@ -1,10 +1,10 @@
 import { Observable } from "rxjs";
 
-import { SendData } from "../../models/data/send.data";
-import { EncArrayBuffer } from "../../models/domain/enc-array-buffer";
-import { Send } from "../../models/domain/send";
-import { SymmetricCryptoKey } from "../../models/domain/symmetric-crypto-key";
-import { SendView } from "../../models/view/send.view";
+import { EncArrayBuffer } from "../../../models/domain/enc-array-buffer";
+import { SymmetricCryptoKey } from "../../../models/domain/symmetric-crypto-key";
+import { SendData } from "../models/data/send.data";
+import { Send } from "../models/domain/send";
+import { SendView } from "../models/view/send.view";
 
 export abstract class SendService {
   sends$: Observable<Send[]>;
