+++ libs/common/src/vault/services/sync/sync.service.ts
@@ -2,8 +2,6 @@ import { ApiService } from "../../../abstractions/api.service";
 import { CryptoService } from "../../../abstractions/crypto.service";
 import { LogService } from "../../../abstractions/log.service";
 import { MessagingService } from "../../../abstractions/messaging.service";
-import { SendApiService } from "../../../abstractions/send/send-api.service.abstraction";
-import { InternalSendService } from "../../../abstractions/send/send.service.abstraction";
 import { SettingsService } from "../../../abstractions/settings.service";
 import { StateService } from "../../../abstractions/state.service";
 import { CollectionService } from "../../../admin-console/abstractions/collection.service";
@@ -18,7 +16,6 @@ import { CollectionDetailsResponse } from "../../../admin-console/models/respons
 import { PolicyResponse } from "../../../admin-console/models/response/policy.response";
 import { KeyConnectorService } from "../../../auth/abstractions/key-connector.service";
 import { sequentialize } from "../../../misc/sequentialize";
-import { SendData } from "../../../models/data/send.data";
 import { DomainsResponse } from "../../../models/response/domains.response";
 import {
   SyncCipherNotification,
@@ -26,7 +23,10 @@ import {
   SyncSendNotification,
 } from "../../../models/response/notification.response";
 import { ProfileResponse } from "../../../models/response/profile.response";
-import { SendResponse } from "../../../models/response/send.response";
+import { SendData } from "../../../tools/send/models/data/send.data";
+import { SendResponse } from "../../../tools/send/models/response/send.response";
+import { SendApiService } from "../../../tools/send/services/send-api.service.abstraction";
+import { InternalSendService } from "../../../tools/send/services/send.service.abstraction";
 import { CipherService } from "../../../vault/abstractions/cipher.service";
 import { FolderApiServiceAbstraction } from "../../../vault/abstractions/folder/folder-api.service.abstraction";
 import { InternalFolderService } from "../../../vault/abstractions/folder/folder.service.abstraction";
